# Conflicts:
#	eFormAPI/eFormAPI.Web/Controllers/Eforms/TemplateFilesController.cs
#	eFormAPI/eFormAPI.Web/Program.cs
#	eFormAPI/eFormAPI.Web/Services/AccountService.cs
#	eFormAPI/eFormAPI.Web/Services/AdminService.cs
#	eFormAPI/eFormAPI.Web/Services/AuthService.cs
#	eFormAPI/eFormAPI.Web/Services/CasesService.cs
#	eFormAPI/eFormAPI.Web/Services/EmailSender.cs
#	eFormAPI/eFormAPI.Web/Services/EntitySearchService.cs
#	eFormAPI/eFormAPI.Web/Services/EntitySelectService.cs
#	eFormAPI/eFormAPI.Web/Services/SimpleSitesService.cs
#	eFormAPI/eFormAPI.Web/Services/SitesService.cs
#	eFormAPI/eFormAPI.Web/Services/TagsService.cs
#	eFormAPI/eFormAPI.Web/Services/TemplateColumnsService.cs
#	eFormAPI/eFormAPI.Web/Services/UnitsService.cs
#	eFormAPI/eFormAPI.Web/Services/WorkersService.cs

+++ eFormAPI/eFormAPI.Web/Services/AdminService.cs
@@ -3,6 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using eFormAPI.Web.Abstractions;
+using eFormAPI.Web.Infrastructure.Database;
+using eFormAPI.Web.Infrastructure.Database.Entities;
+using eFormAPI.Web.Infrastructure.Models.Users;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
@@ -11,7 +14,6 @@ using Microting.eFormApi.BasePn.Infrastructure.Helpers.WritableOptions;
 using Microting.eFormApi.BasePn.Infrastructure.Models.Application;
 using Microting.eFormApi.BasePn.Infrastructure.Models.API;
 using Microting.eFormApi.BasePn.Infrastructure.Models.Common;
-using Microting.eFormApi.BasePn.Infrastructure.Models.User;
 
 namespace eFormAPI.Web.Services
 {
@@ -21,33 +23,36 @@ namespace eFormAPI.Web.Services
         private readonly IWritableOptions<ApplicationSettings> _appSettings;
         private readonly ILogger<AdminService> _logger;
         private readonly ILocalizationService _localizationService;
+        private readonly BaseDbContext _dbContext;
         private readonly UserManager<EformUser> _userManager;
 
         public AdminService(ILogger<AdminService> logger,
             UserManager<EformUser> userManager,
             IWritableOptions<ApplicationSettings> appSettings,
-            IUserService userService, 
-            ILocalizationService localizationService)
+            IUserService userService,
+            ILocalizationService localizationService,
+            BaseDbContext dbContext)
         {
             _logger = logger;
             _userManager = userManager;
             _appSettings = appSettings;
             _userService = userService;
             _localizationService = localizationService;
+            _dbContext = dbContext;
         }
 
         public async Task<OperationDataResult<UserRegisterModel>> GetUser(int userId)
         {
             try
             {
-                EformUser user = await _userService.GetByIdAsync(userId);
+                var user = await _userService.GetByIdAsync(userId);
                 if (user == null)
                 {
                     return new OperationDataResult<UserRegisterModel>(false,
                         _localizationService.GetString("UserNotFound"));
                 }
 
-                UserRegisterModel result = new UserRegisterModel()
+                var result = new UserRegisterModel()
                 {
                     Email = user.Email,
                     Id = user.Id,
@@ -56,8 +61,14 @@ namespace eFormAPI.Web.Services
                     UserName = user.UserName,
                 };
                 // get role
-                IList<string> roles = await _userManager.GetRolesAsync(user);
+                var roles = await _userManager.GetRolesAsync(user);
                 result.Role = roles.FirstOrDefault();
+                // get user group
+                result.GroupId = await _dbContext.SecurityGroupUsers
+                    .Where(x => x.EformUserId == user.Id)
+                    .Select(x => x.SecurityGroup.Id)
+                    .FirstOrDefaultAsync();
+
                 return new OperationDataResult<UserRegisterModel>(true, result);
             }
             catch (Exception exception)
@@ -72,8 +83,8 @@ namespace eFormAPI.Web.Services
         {
             try
             {
-                List<UserInfoViewModel> userList = new List<UserInfoViewModel>();
-                List<EformUser> userResult = _userManager.Users
+                var userList = new List<UserInfoViewModel>();
+                var userResult = _userManager.Users
                     .Include(x => x.UserRoles)
                     .ThenInclude(x => x.Role)
                     .OrderBy(z => z.Id)
@@ -83,8 +94,8 @@ namespace eFormAPI.Web.Services
 
                 userResult.ForEach(userItem =>
                 {
-                    string roleName = userItem.UserRoles.Select(y => y.Role.Name).FirstOrDefault();
-                    UserInfoViewModel modelItem = new UserInfoViewModel();
+                    var roleName = userItem.UserRoles.Select(y => y.Role.Name).FirstOrDefault();
+                    var modelItem = new UserInfoViewModel();
                     if (roleName != null)
                     {
                         modelItem.Role = roleName;
@@ -97,7 +108,7 @@ namespace eFormAPI.Web.Services
                     modelItem.UserName = userItem.UserName;
                     userList.Add(modelItem);
                 });
-                int totalUsers = _userManager.Users.Count();
+                var totalUsers = _userManager.Users.Count();
                 return new OperationDataResult<UserInfoModelList>(true, new UserInfoModelList()
                 {
                     TotalUsers = totalUsers,
@@ -116,11 +127,17 @@ namespace eFormAPI.Web.Services
         {
             try
             {
-                EformUser user = await _userService.GetByIdAsync(userRegisterModel.Id);
+                var user = await _userService.GetByIdAsync(userRegisterModel.Id);
                 if (user == null)
                 {
                     return new OperationResult(false,
                         _localizationService.GetString("UserNotFoundUserName", userRegisterModel.UserName));
+                }
+
+                if (!_dbContext.SecurityGroups.Any(x => x.Id == userRegisterModel.GroupId))
+                {
+                    return new OperationResult(false,
+                        _localizationService.GetString("SecurityGroupNotFound"));
                 }
 
                 if (userRegisterModel.Role == null)
@@ -128,11 +145,17 @@ namespace eFormAPI.Web.Services
                     return new OperationResult(false, _localizationService.GetString("RoleIsRequired"));
                 }
 
+                if (await _userManager.IsInRoleAsync(user, EformRole.Admin)
+                    && _userService.Role != EformRole.Admin)
+                {
+                    return new OperationResult(false, _localizationService.GetString("YouCantViewChangeOrDeleteAdmin"));
+                }
+
                 user.Email = userRegisterModel.Email;
                 user.UserName = userRegisterModel.UserName;
                 user.FirstName = userRegisterModel.FirstName;
                 user.LastName = userRegisterModel.LastName;
-                IdentityResult result = await _userManager.UpdateAsync(user);
+                var result = await _userManager.UpdateAsync(user);
                 if (!result.Succeeded)
                 {
                     return new OperationResult(false, string.Join(" ", result.Errors));
@@ -145,10 +168,30 @@ namespace eFormAPI.Web.Services
                     await _userManager.AddPasswordAsync(user, userRegisterModel.Password);
                 }
 
-                // change role
-                await _userManager.RemoveFromRolesAsync(user, new[] {EformRole.Admin, EformRole.User});
-                await _userManager.AddToRoleAsync(user, userRegisterModel.Role);
-                return new OperationResult(true, _localizationService.GetString("UserUserNameWasUpdated", user.UserName));
+                // Change group
+                if (userRegisterModel.GroupId > 0 && user.Id > 0)
+                {
+                    var securityGroupUsers = _dbContext.SecurityGroupUsers
+                        .Where(x => x.EformUserId == user.Id
+                                    && x.SecurityGroupId != userRegisterModel.GroupId);
+
+                    _dbContext.SecurityGroupUsers.RemoveRange(securityGroupUsers);
+                    if (!_dbContext.SecurityGroupUsers.Any(x =>
+                        x.EformUserId == user.Id && x.SecurityGroupId == userRegisterModel.GroupId))
+                    {
+                        var securityGroupUser = new SecurityGroupUser()
+                        {
+                            SecurityGroupId = (int) userRegisterModel.GroupId,
+                            EformUserId = user.Id
+                        };
+                        _dbContext.SecurityGroupUsers.Add(securityGroupUser);
+                    }
+
+                    await _dbContext.SaveChangesAsync();
+                }
+
+                return new OperationResult(true,
+                    _localizationService.GetString("UserUserNameWasUpdated", user.UserName));
             }
             catch (Exception exception)
             {
@@ -161,19 +204,20 @@ namespace eFormAPI.Web.Services
         {
             try
             {
-                EformUser userResult = await _userManager.FindByNameAsync(userRegisterModel.UserName);
+                var userResult = await _userManager.FindByNameAsync(userRegisterModel.UserName);
                 if (userResult != null)
                 {
                     return new OperationResult(false,
                         _localizationService.GetString("UserUserNameAlreadyExist", userRegisterModel.UserName));
                 }
 
-                if (userRegisterModel.Role == null)
+                if (!_dbContext.SecurityGroups.Any(x => x.Id == userRegisterModel.GroupId))
                 {
-                    return new OperationResult(false, _localizationService.GetString("RoleIsRequired"));
+                    return new OperationResult(false,
+                        _localizationService.GetString("SecurityGroupNotFound"));
                 }
 
-                EformUser user = new EformUser
+                var user = new EformUser
                 {
                     Email = userRegisterModel.Email,
                     UserName = userRegisterModel.UserName,
@@ -183,15 +227,28 @@ namespace eFormAPI.Web.Services
                     IsGoogleAuthenticatorEnabled = false
                 };
 
-                IdentityResult result = await _userManager.CreateAsync(user, userRegisterModel.Password);
+                var result = await _userManager.CreateAsync(user, userRegisterModel.Password);
                 if (!result.Succeeded)
                 {
                     return new OperationResult(false, string.Join(" ", result.Errors));
                 }
 
                 // change role
-                await _userManager.AddToRoleAsync(user, userRegisterModel.Role.ToLower());
-                return new OperationResult(true, _localizationService.GetString("UserUserNameWasCreated", user.UserName));
+                await _userManager.AddToRoleAsync(user, EformRole.User);
+                // add to group
+                if (userRegisterModel.GroupId > 0 && user.Id > 0)
+                {
+                    var securityGroupUser = new SecurityGroupUser()
+                    {
+                        SecurityGroupId = (int) userRegisterModel.GroupId,
+                        EformUserId = user.Id
+                    };
+                    _dbContext.SecurityGroupUsers.Add(securityGroupUser);
+                    await _dbContext.SaveChangesAsync();
+                }
+
+                return new OperationResult(true,
+                    _localizationService.GetString("UserUserNameWasCreated", user.UserName));
             }
             catch (Exception exception)
             {
@@ -209,13 +266,19 @@ namespace eFormAPI.Web.Services
                     return new OperationResult(false, _localizationService.GetString("CantDeletePrimaryAdminUser"));
                 }
 
-                EformUser user = await _userService.GetByIdAsync(userId);
+                var user = await _userService.GetByIdAsync(userId);
+                if (await _userManager.IsInRoleAsync(user, EformRole.Admin)
+                    && _userService.Role != EformRole.Admin)
+                {
+                    return new OperationResult(false, _localizationService.GetString("YouCantViewChangeOrDeleteAdmin"));
+                }
+
                 if (user == null)
                 {
                     return new OperationResult(false, _localizationService.GetString("UserUserNameNotFound", userId));
                 }
 
-                IdentityResult result = await _userManager.DeleteAsync(user);
+                var result = await _userManager.DeleteAsync(user);
                 if (!result.Succeeded)
                 {
                     return new OperationResult(false, string.Join(" ", result.Errors));
